fix(frontend): replace empty /community route with the working implementation
Same failure pattern as insurance: app/[locale]/community/page.tsx (the
only reachable copy) was a completely empty file — 0 bytes, no export at
all — which crashes Next.js during static export with "Unsupported Server
Component type" for every locale, not just a type error.

app/community/page.tsx had the real, working, self-contained page (plain
Tailwind markup + the existing CommunityLinks component), just stranded
at an unreachable URL by the i18n migration. Moved it into the reachable
[locale] slot and removed the now-redundant unreachable copy.
